Drivers: deprecate unused member functions
1) VelodyneParser::get_nan_point
2) VelodyneDriver::SetBaseTime

+++ modules/drivers/lidar/velodyne/driver/socket_input.cc
@@ -59,7 +59,7 @@ void SocketInput::init(const int &port) {
   }
 
   // connect to Velodyne UDP port
-  AINFO << "Opening UDP socket: port " << uint16_t(port);
+  AINFO << "Opening UDP socket: port " << port;
   port_ = port;
   sockfd_ = socket(AF_INET, SOCK_DGRAM, 0);
